[chore] tui fixed
